demo-gitlab-api/requirement.txt
numpy==1.19.1
openpyxl==3.0.4
pandas==1.0.5
python-gitlab==2.4.0
requests==2.24.0
six==1.15.0
